fix(cli): Harden remote config review against symlinked configs
- Refuse a remote config that is a symlink or whose real path escapes the clone.
  Git preserves symlinks on checkout and config resolution followed them, so the
  reviewed bytes could come from outside the owner-only temp dir: the prompt could
  echo an unrelated local file, and the target could be swapped between the review
  read and the later load.
- Report an unreadable config as a RepomixError with context instead of letting the
  raw fs error escape.
- Correct the content-pinning claim. The digest covers the entry config file only,
  so a .ts/.js config can import modules that change without re-prompting. Say so
  in the store comment and in the docs across all 15 language directories, rather
  than implying the pin covers everything the config runs.
- Split the trust store into remoteConfigTrustStore.ts. The two halves are
  security-sensitive on-disk state and interactive UI, and only meet through two
  exported functions.
