+++ website/client/src/hi/guide/remote-repository-processing.md
@@ -93,7 +93,7 @@ REPOMIX_REMOTE_TRUST_CONFIG=true repomix --remote user/repo
 जब आप किसी इंटरैक्टिव टर्मिनल में किसी रिपॉजिटरी के कॉन्फ़िग पर भरोसा करते हैं, तो repomix उस कॉन्फ़िग को दिखाता है जो चलने वाला है और उसे लोड करने से पहले आपसे पुष्टि माँगता है:
 
 - **हाँ, सिर्फ़ इस बार**: केवल इस रन पर भरोसा करें।
-- **हाँ, और इस रिपॉजिटरी के लिए दोबारा न पूछें**: यह तब तक याद रखा जाता है जब तक आपकी अस्थायी फ़ाइलें साफ़ नहीं हो जातीं, और केवल तब तक जब तक कॉन्फ़िग वही रहता है (कॉन्फ़िग बदलने पर फिर से पूछा जाएगा)।
+- **हाँ, और इस रिपॉजिटरी के लिए दोबारा न पूछें**: यह तब तक याद रखा जाता है जब तक आपकी अस्थायी फ़ाइलें साफ़ नहीं हो जातीं, और केवल तब तक जब तक वह कॉन्फ़िग फ़ाइल नहीं बदलती (कॉन्फ़िग फ़ाइल में बदलाव होने पर फिर से पूछा जाएगा)। ध्यान दें कि यह जाँच केवल कॉन्फ़िग फ़ाइल पर ही लागू होती है: एक `.ts` / `.js` कॉन्फ़िग अन्य फ़ाइलों को इम्पोर्ट कर सकता है, और वे इस जाँच का हिस्सा नहीं हैं।
 - **नहीं**: कॉन्फ़िग को चलाए बिना रोक दें।
 
 `--force` पास करने पर, CI जैसे गैर-इंटरैक्टिव शेल में (कॉन्फ़िग पहले की तरह भरोसेमंद माना जाता है, जिससे मौजूदा ऑटोमेशन काम करता रहता है), या जब आप पहले ही उस रिपॉजिटरी पर हमेशा भरोसा करने का चुनाव कर चुके हों, तो यह प्रॉम्प्ट छोड़ दिया जाता है।
