+++ website/client/src/ja/guide/remote-repository-processing.md
@@ -93,7 +93,7 @@ REPOMIX_REMOTE_TRUST_CONFIG=true repomix --remote user/repo
 インタラクティブなターミナルでリポジトリの設定を信頼すると、repomixはこれから実行しようとしている設定を表示し、読み込む前に確認を求めます。
 
 - **はい、今回のみ**：この実行だけを信頼します。
-- **はい、このリポジトリでは今後確認しない**：一時ファイルが削除されるまで、かつ設定が変更されない限り記憶されます（設定が変更されると再度確認が求められます）。
+- **はい、このリポジトリでは今後確認しない**：一時ファイルが削除されるまで、かつその設定ファイル自体が変更されない限り記憶されます（設定ファイルが編集されると再度確認が求められます）。なお、この確認の対象は設定ファイル自体のみです。`.ts` / `.js` 形式の設定は他のファイルをインポートできますが、それらはチェックの対象に含まれません。
 - **いいえ**：設定を実行せずに中止します。
 
 `--force` を指定した場合、CIのような非インタラクティブなシェルの場合（これまでどおり設定は信頼され、既存の自動化はそのまま動作します）、またはそのリポジトリを常に信頼するとすでに選択している場合は、この確認プロンプトはスキップされます。
