+++ website/client/src/ko/guide/remote-repository-processing.md
@@ -93,7 +93,7 @@ REPOMIX_REMOTE_TRUST_CONFIG=true repomix --remote user/repo
 대화형 터미널에서 저장소의 설정을 신뢰하면, repomix는 실행하려는 설정을 보여주고 불러오기 전에 확인을 요청합니다.
 
 - **예, 이번만**: 이번 실행만 신뢰합니다.
-- **예, 이 저장소에서는 다시 묻지 않음**: 임시 파일이 삭제될 때까지, 그리고 설정이 동일하게 유지되는 동안에만 기억됩니다(설정이 변경되면 다시 확인을 요청합니다).
+- **예, 이 저장소에서는 다시 묻지 않음**: 임시 파일이 삭제될 때까지, 그리고 해당 설정 파일이 변경되지 않는 동안에만 기억됩니다(설정 파일이 수정되면 다시 확인을 요청합니다). 이 확인은 설정 파일 자체에만 적용된다는 점에 유의하세요. `.ts` / `.js` 설정은 다른 파일을 가져올 수 있으며, 그런 파일은 이 검사 대상에 포함되지 않습니다.
 - **아니요**: 설정을 실행하지 않고 중단합니다.
 
 `--force`를 전달했거나, CI와 같은 비대화형 셸을 사용하거나(설정은 이전과 마찬가지로 신뢰되어 기존 자동화가 계속 작동합니다), 이미 해당 저장소를 항상 신뢰하도록 선택한 경우에는 이 확인 프롬프트가 표시되지 않습니다.
