+++ website/client/src/ru/guide/remote-repository-processing.md
@@ -93,7 +93,7 @@ REPOMIX_REMOTE_TRUST_CONFIG=true repomix --remote user/repo
 Когда вы доверяете конфигурации репозитория в интерактивном терминале, repomix показывает конфигурацию, которая будет выполнена, и просит подтвердить перед её загрузкой:
 
 - **Да, только один раз**: доверять только этому запуску.
-- **Да, и больше не спрашивать для этого репозитория**: запоминается до очистки временных файлов и только пока конфигурация остаётся неизменной (при изменении конфигурации запрос появится снова).
+- **Да, и больше не спрашивать для этого репозитория**: запоминается до очистки временных файлов и только пока этот файл конфигурации остаётся неизменным (при изменении файла конфигурации запрос появится снова). Обратите внимание, что проверка распространяется только на сам файл конфигурации: конфигурация `.ts` / `.js` может импортировать другие файлы, и они не учитываются при проверке.
 - **Нет**: прервать без выполнения конфигурации.
 
 Запрос пропускается, если вы передали `--force`, в неинтерактивных оболочках, таких как CI (конфигурация считается доверенной как и раньше, что сохраняет работу существующей автоматизации), или если вы уже выбрали всегда доверять этому репозиторию.
