+++ website/client/src/vi/guide/remote-repository-processing.md
@@ -93,7 +93,7 @@ REPOMIX_REMOTE_TRUST_CONFIG=true repomix --remote user/repo
 Khi bạn tin tưởng cấu hình của một kho lưu trữ trong terminal tương tác, repomix sẽ hiển thị cấu hình sắp được chạy và yêu cầu bạn xác nhận trước khi tải nó:
 
 - **Có, chỉ lần này**: chỉ tin tưởng lần chạy này.
-- **Có, và không hỏi lại cho kho lưu trữ này**: được ghi nhớ cho đến khi các tệp tạm thời của bạn bị xóa, và chỉ khi cấu hình không thay đổi (cấu hình thay đổi sẽ khiến lời nhắc xuất hiện lại).
+- **Có, và không hỏi lại cho kho lưu trữ này**: được ghi nhớ cho đến khi các tệp tạm thời của bạn bị xóa, và chỉ khi tệp cấu hình đó không thay đổi (tệp cấu hình bị chỉnh sửa sẽ khiến lời nhắc xuất hiện lại). Lưu ý rằng điều này chỉ áp dụng cho chính tệp cấu hình: một cấu hình `.ts` / `.js` có thể import các tệp khác, và những tệp đó không nằm trong phạm vi kiểm tra này.
 - **Không**: hủy bỏ mà không chạy cấu hình.
 
 Lời nhắc này sẽ được bỏ qua khi bạn truyền `--force`, trong các shell không tương tác như CI (cấu hình vẫn được tin tưởng như trước, giúp các quy trình tự động hiện có tiếp tục hoạt động), hoặc khi bạn đã chọn luôn tin tưởng kho lưu trữ đó.
