+++ website/client/src/zh-cn/guide/remote-repository-processing.md
@@ -93,7 +93,7 @@ REPOMIX_REMOTE_TRUST_CONFIG=true repomix --remote user/repo
 当你在交互式终端中信任某个仓库的配置时，repomix 会显示即将运行的配置，并在加载前要求你确认：
 
 - **仅本次同意**：仅信任这一次运行。
-- **同意，且不再询问此仓库**：会一直记住，直到清除临时文件为止，且仅在配置保持不变时有效（配置发生变化会再次提示）。
+- **同意，且不再询问此仓库**：会一直记住，直到清除临时文件为止，且仅在该配置文件保持不变时有效（配置文件被修改后会再次提示）。请注意，此检查仅针对配置文件本身：`.ts` / `.js` 配置可以导入其他文件，这些文件不在检查范围内。
 - **否**：中止操作，不运行该配置。
 
 当你传入 `--force`、在 CI 等非交互式 shell 中运行（配置会像以前一样被信任，从而保持现有自动化正常工作），或者你已经选择始终信任该仓库时，将跳过此提示。
