+++ website/client/src/zh-tw/guide/remote-repository-processing.md
@@ -93,7 +93,7 @@ REPOMIX_REMOTE_TRUST_CONFIG=true repomix --remote user/repo
 當您在互動式終端機中信任某個倉庫的設定時，repomix 會顯示即將執行的設定，並在載入前要求您確認：
 
 - **僅本次同意**：僅信任這一次執行。
-- **同意，且不再詢問此倉庫**：會持續記住，直到清除暫存檔案為止，且僅在設定保持不變時有效（設定變更後會再次提示）。
+- **同意，且不再詢問此倉庫**：會持續記住，直到清除暫存檔案為止，且僅在該設定檔保持不變時有效（設定檔被修改後會再次提示）。請注意，此檢查僅針對設定檔本身：`.ts` / `.js` 設定可以匯入其他檔案，這些檔案不在檢查範圍內。
 - **否**：中止操作，不執行該設定。
 
 當您傳入 `--force`、在 CI 等非互動式 shell 中執行（設定會如以往一樣被信任，讓現有自動化保持正常運作），或是您已經選擇一律信任該倉庫時，此提示將被跳過。
